pyproject.toml
[project]
name = "votedge"
version = "0.1.0"
description = "Political data collection and analysis for election prediction"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "GoogleNews",
    "requests",
    "tweepy",
    "python-dotenv",
    "pandas",
    "textblob",
    "matplotlib",
    "seaborn",
    "playwright"
]

[project.scripts]
votedge = "votedge.main:main"

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"